requirements.txt
anyio==3.6.1
click==8.1.3
dnspython==2.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
pymongo==4.1.1
python-decouple==3.6
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
uvicorn==0.18.2
